requirements.txt
requests
sqlalchemy==1.4.40
psycopg2
python-dotenv